requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.2.1
httpx>=0.26.0
Pillow>=10.0.0
pdf2image>=1.16.3